Fix #612 (#617)
* Fix #612

* Outsource genome search box test

* Move genome search box test to the corresponding file

* Fix lookup issue
